LRAC-7374 expecting Integer

+++ modules/dxp/apps/osb/osb-asah/osb-asah-extractor/src/main/java/com/liferay/osb/asah/extractor/processor/DXPEntitiesMessageProcessor.java
@@ -292,7 +292,7 @@ public class DXPEntitiesMessageProcessor {
 					put(
 						"dataSourceId",
 						objectJSONObject.getString("osbAsahDataSourceId"));
-					put("userId", objectJSONObject.getLong("userId"));
+					put("userId", objectJSONObject.getInt("userId"));
 				}
 			},
 			DXPEntity.Type.USER);
